worker/pyproject.toml
[project]
name = "whiteops-worker"
version = "0.1.0"
description = "White-Ops Agent Worker"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    "litellm>=1.50.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    # Office tools
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "pypdf>=5.1.0",
    "reportlab>=4.2.0",
    # Browser & web
    "playwright>=1.49.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    # Data & analysis
    "matplotlib>=3.9.0",
    "plotly>=5.24.0",
    # Image
    "Pillow>=11.0.0",
    # File storage
    "minio>=7.2.0",
    # Email
    "aiosmtplib>=3.0.0",
    "aioimaplib>=2.0.0",
    # Misc
    "psutil>=6.1.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
